Split up compilation, linking and execution into separate steps.
Previously, we rolled these all together in `compileAndEval`, while in other
places we separated them, like in compiling top-level functions. This had a few
problems:
  1. We ended up with different compilation paths for top-level functions,
     evaluated expressions, and export/foreign functions.
  2. Long-running linking meant a long-running compilation jit, which led to
     concerns about thread safety of LLVM APIs (see Adam's comments on PR
     #1091). Long-running state also makes caching tricky and casts shadows in
     lots of places.
  3. We exposed a lot of LLVM surface area because of all the different ways in
     which we interacted with it.

Main changes
  * All compilation now goes through `LLVM.Compile`, which produces object code
  * All linking now goes through `LLVM.Link`, which yields a pointer to
    executable code and a callback for cleanup. We don't even need to use LLVM
    for this, and we could swap it out for `ld` or something.
  * All execution goes through `Runtime`, which doesn't call into LLVM at all!
  * There is no persistent LLVM state at all. We create a temporary jit every
    time we need to compile a function and we destroy it once we have the object
    code. We create separate temporary jits each time we need to link/load some
    object code. We keep these jits alive until we destroy them by calling the
    teardown callback for the linked function. (We might want to be more
    thoughtful about how we manage resources for these.)
  * `Dex.Foreign` stores the top env inside an `MVar` rather than passing it
    around as a (pointer to a) value. This takes care of threading issues
    (because MVars work as a lock) and ensures that we don't accidentally use
    the wrong env.
  * Renamed the imp-to-llvm translation pass to `ImpToLLVM` instead of `JIT` (we
    had so many files called `JIT.hs`. In this change, we remove three of them!)
