+++ src/Dex/Foreign/Context.hs
@@ -7,19 +7,21 @@
 {-# LANGUAGE RankNTypes #-}
 
 module Dex.Foreign.Context (
-  Context (..), AtomEx (..),
+  Context (..), AtomEx (..), ExportNativeFunctionAddr, ExportNativeFunction (..),
+  ClosedExportedSignature,
   setError, runTopperMFromContext,
   dexCreateContext, dexDestroyContext, dexForkContext,
   dexInsert, dexLookup,
-  dexEval, dexFreshName,
+  dexEval, dexFreshName, insertIntoNativeFunctionTable
   ) where
 
+import Foreign.C
 import Foreign.Ptr
 import Foreign.StablePtr
-import Foreign.C.String
 import System.Random
 
 import Control.Monad.IO.Class
+import Control.Concurrent.MVar
 import Data.Functor
 import Data.Foldable
 import Data.Map.Strict    qualified as M
@@ -33,62 +35,80 @@ import PPrint
 import Err
 import AbstractSyntax
 import Builder
+import Export
 
 import Dex.Foreign.Util
 
+data Context = Context EvalConfig (MVar TopStateEx) (MVar NativeFunctionTable)
 
-data Context = Context EvalConfig DexJIT TopStateEx
+type ClosedExportedSignature = ExportedSignature 'VoidS
+data ExportNativeFunction =
+  ExportNativeFunction { nativeFunction  :: NativeFunction
+                       , nativeSignature :: ExportedSignature 'VoidS }
+
+type ExportNativeFunctionAddr = Ptr () -- points to executable code
+type NativeFunctionTable = M.Map ExportNativeFunctionAddr ExportNativeFunction
+
 data AtomEx where
   AtomEx :: Atom n -> AtomEx
 
 dexCreateContext :: IO (Ptr Context)
 dexCreateContext = do
-  let evalConfig = EvalConfig LLVM [LibBuiltinPath] Nothing Nothing Nothing Optimize
+  let cfg = EvalConfig LLVM [LibBuiltinPath] Nothing Nothing Nothing Optimize
   cachedEnv <- loadCache
-  jit <- createDexJIT (getLLVMOptLevel evalConfig)
-  runTopperM evalConfig jit cachedEnv (evalSourceBlockRepl preludeImportBlock) >>= \case
-    (Result _  (Success  ()), preludeEnv) -> toStablePtr $ Context evalConfig jit preludeEnv
-    (Result _  (Failure err), _         ) -> nullPtr <$
-      setError ("Failed to initialize standard library: " ++ pprint err)
+  envMVar      <- newMVar cachedEnv
+  ptrTableMVar <- newMVar mempty
+  ctxPtr <- toStablePtr $ Context cfg envMVar ptrTableMVar
+  _ <- runTopperMFromContext ctxPtr $ evalSourceBlockRepl preludeImportBlock
+  return ctxPtr
 
 dexDestroyContext :: Ptr Context -> IO ()
 dexDestroyContext = freeStablePtr . castPtrToStablePtr . castPtr
 
 dexForkContext :: Ptr Context -> IO (Ptr Context)
-dexForkContext ctxPtr = toStablePtr =<< fromStablePtr ctxPtr
+dexForkContext ctxPtr = do
+  Context cfg envMVar ptrTabMVar <- fromStablePtr ctxPtr
+  envMVar'    <- copyMVar envMVar
+  ptrTabMVar' <- copyMVar ptrTabMVar
+  toStablePtr $ Context cfg envMVar' ptrTabMVar'
 
-runTopperMFromContext :: Ptr Context -> (forall n. Mut n => TopperM n a) -> IO (a, TopStateEx)
+-- XXX: this holds the lock on the mvar for the whole duration. TODO: consider
+-- implementing a new instance of `Topper`, backed by an mvar, that only holds
+-- the lock when it actually needs to update the env.
+runTopperMFromContext :: Ptr Context -> (forall n. Mut n => TopperM n a) -> IO a
 runTopperMFromContext ctxPtr cont = do
-  Context evalConfig jit env <- fromStablePtr ctxPtr
-  runTopperM evalConfig jit env cont
+  Context cfg envMVar _ <- fromStablePtr ctxPtr
+  env <- takeMVar envMVar
+  (result, newEnv) <- runTopperM cfg env cont
+  putMVar envMVar newEnv
+  return result
 {-# INLINE runTopperMFromContext #-}
 
-dexEval :: Ptr Context -> CString -> IO (Ptr Context)
+dexEval :: Ptr Context -> CString -> IO CInt
 dexEval ctxPtr sourcePtr = do
   source <- T.decodeUtf8 <$> BS.packCString sourcePtr
-  (results, finalEnv) <- runTopperMFromContext ctxPtr $
+  Context cfg envMVar _ <- fromStablePtr ctxPtr
+  oldEnv <- takeMVar envMVar
+  (results, newEnv) <- runTopperM cfg oldEnv $
     evalSourceText source (const $ return ()) (const $ return True)
   let anyError = asum $ fmap (\case (_, Result _ (Failure err)) -> Just err; _ -> Nothing) results
-  Context evalConfig jit _ <- fromStablePtr ctxPtr
   case anyError of
-    Nothing  -> toStablePtr $ Context evalConfig jit finalEnv
-    Just err -> setError (pprint err) $> nullPtr
+    Nothing  -> putMVar envMVar newEnv                          >> return 0
+    Just err -> putMVar envMVar oldEnv >> setError (pprint err) >> return 1
 
-dexInsert :: Ptr Context -> CString -> Ptr AtomEx -> IO (Ptr Context)
+dexInsert :: Ptr Context -> CString -> Ptr AtomEx -> IO ()
 dexInsert ctxPtr namePtr atomPtr = do
   sourceName <- peekCString namePtr
   AtomEx atom <- fromStablePtr atomPtr
-  (_, finalEnv) <- runTopperMFromContext ctxPtr do
+  runTopperMFromContext ctxPtr do
     -- TODO: Check if atom is compatible with context! Use module name?
     name <- emitTopLet (getNameHint @String sourceName) PlainLet $ Atom $ unsafeCoerceE atom
     emitSourceMap $ SourceMap $ M.singleton sourceName [ModuleVar Main $ Just $ UAtomVar name]
-  Context evalConfig jit _ <- fromStablePtr ctxPtr
-  toStablePtr $ Context evalConfig jit finalEnv
 
 dexLookup :: Ptr Context -> CString -> IO (Ptr AtomEx)
 dexLookup ctxPtr namePtr = do
   name <- peekCString namePtr
-  fst <$> runTopperMFromContext ctxPtr do
+  runTopperMFromContext ctxPtr do
     lookupSourceMap name >>= \case
       Just (UAtomVar v) -> lookupAtomName v >>= \case
         LetBound (DeclBinding _ _ (Atom atom)) -> liftIO $ toStablePtr $ AtomEx atom
@@ -98,8 +118,9 @@ dexLookup ctxPtr namePtr = do
 
 dexFreshName :: Ptr Context -> IO CString
 dexFreshName ctxPtr = do
-  Context evalConfig jit env <- fromStablePtr ctxPtr
-  (cstr, _) <- runTopperM evalConfig jit env genName
+  Context evalConfig envMVar _  <- fromStablePtr ctxPtr
+  env <- readMVar envMVar
+  (cstr, _) <- runTopperM evalConfig env genName
   return cstr
   where
     genName :: Topper m => m n CString
@@ -110,3 +131,10 @@ dexFreshName ctxPtr = do
       lookupSourceMap name >>= \case
         Nothing -> liftIO $ newCString name
         Just _  -> genName
+
+insertIntoNativeFunctionTable
+  :: Ptr Context -> ExportNativeFunctionAddr -> ExportNativeFunction -> IO ()
+insertIntoNativeFunctionTable ctxPtr funPtr funData = do
+  Context _ _ addrTableMVar  <- fromStablePtr ctxPtr
+  modifyMVar addrTableMVar (\m -> return (M.insert funPtr funData m, ()))
+
