+++ src/lib/Core.hs
@@ -408,8 +408,8 @@ lookupModule :: EnvReader m => ModuleName n -> m n (Module n)
 lookupModule name = lookupEnv name >>= \case ModuleBinding m -> return m
 {-# INLINE lookupModule #-}
 
-lookupFunObjCode :: EnvReader m => FunObjCodeName n -> m n (FunObjCode n)
-lookupFunObjCode name = lookupEnv name >>= \case FunObjCodeBinding m -> return m
+lookupFunObjCode :: EnvReader m => FunObjCodeName n -> m n (FunObjCode, FunObjCodeNameMap n)
+lookupFunObjCode name = lookupEnv name >>= \case FunObjCodeBinding obj m -> return (obj, m)
 {-# INLINE lookupFunObjCode #-}
 
 lookupDataDef :: EnvReader m => DataDefName n -> m n (DataDef n)
