--- src/lib/LLVM/JIT.hs
@@ -1,13 +1,0 @@
--- Copyright 2021 Google LLC
---
--- Use of this source code is governed by a BSD-style
--- license that can be found in the LICENSE file or at
--- https://developers.google.com/open-source/licenses/bsd
-
-module LLVM.JIT (
-  JIT, createJIT, destroyJIT, withJIT,
-  NativeModule, compileModule, loadNativeModule, unloadNativeModule, withNativeModule,
-  getFunctionPtr
-  ) where
-
-import LLVM.HEAD.JIT
