+++ src/lib/Types/Source.hs
@@ -349,6 +349,8 @@ data LogLevel = LogNothing | PrintEvalTime | PrintBench String
 
 data OutFormat = Printed | RenderHtml  deriving (Show, Eq, Generic)
 
+data OptLevel = NoOptimize | Optimize
+
 data PassName = Parse | RenamePass | TypePass | SynthPass | SimpPass | ImpPass | JitPass
               | LLVMOpt | AsmPass | JAXPass | JAXSimpPass | LLVMEval | LowerOptPass | LowerPass
               | ResultPass | JaxprAndHLO | EarlyOptPass | OptPass
